project_predelivery_1

+++ src/libs/classes/Container.ts
@@ -1,67 +1,89 @@
 import {
   cliError, 
   cliNotice, 
-  cliSuccess 
+  cliSuccess,
+  getTimestamp
 } from "../functions/index.js";
 import * as fs from "fs";
 
 export default class Container {
   _fileDir: string
+  _fields: string[]
 
   constructor (
-    fileDir: string
+    fileDir: string,
+    fields: string[]
   ) {
     this._fileDir = fileDir;
-    this.readOrCreateFile();
+    this._fields = fields;
+    this._readOrCreateFile();
   }
 
   /* ------------------------------------------ */
   /*                   Methods                  */
   /* ------------------------------------------ */
 
-  /* --------- Begin readOrCreateFile --------- */
-  async readOrCreateFile (): Promise<void> {
+  /* --------- Begin _readOrCreateFile --------- */
+  async _readOrCreateFile (): Promise<void> {
     try {
       await fs.promises.readFile(this._fileDir,'utf8')
     } catch (err: any) {
       if (err['code'] === 'ENOENT'){
         cliNotice(`File ${this._fileDir} does not exists\n${err['message']}`);
-        this.createFile();
+        this._createFile();
       } else {
         cliError(`Error Code: ${err['code']} | There was an unexpected error when trying to read ${this._fileDir}\n${err['message']}`);
       }
     } 
   }
-  /* ---------- End readOrCreateFile ---------- */
+  /* ---------- End _readOrCreateFile ---------- */
 
-  /* ------------ Begin createFile ------------ */
-  async createFile (): Promise<void> {
+  /* ------------ Begin _createFile ------------ */
+  async _createFile (): Promise<void> {
     await fs.promises.writeFile(this._fileDir,'[]','utf8')
     cliSuccess(`New file ${this._fileDir} created`);
   }
-  /* ------------- End createFile ------------- */
+  /* ------------- End _createFile ------------- */
+
+  /* ------------- Begin keysExist ------------ */
+  _fieldsExist (data: object): boolean {
+    const dataKeys: string[] = Object.keys(data);
+    let fieldsExist = true;
+    dataKeys.forEach((key: string)=>{
+      if (!this._fields.includes(key)){
+        fieldsExist = false;
+      }
+    })
+    return fieldsExist
+  }
+  /* -------------- End keysExist ------------- */
 
   /* --------------- Begin save --------------- */
   async save(data: object): Promise<number|undefined> {
-    try {
-      const jsonData: Array<object> = JSON.parse(
-        await fs.promises.readFile(this._fileDir,'utf8')
-      );
-      const newData: any = data;
-      if (jsonData.length) {
-        const lastItem: any = jsonData[jsonData.length - 1];
-        const lastId: any | undefined = lastItem['id'] ? lastItem['id'] : undefined;
-        newData["id"] = lastId ? lastId + 1: 1;
-      } else {
-        newData["id"] = 1;
+    if (this._fieldsExist(data)) {
+      try {
+        const jsonData: Array<object> = JSON.parse(
+          await fs.promises.readFile(this._fileDir,'utf8')
+        );
+        const newData: any = data;
+        if (jsonData.length) {
+          const lastItem: any = jsonData[jsonData.length - 1];
+          const lastId: any | undefined = lastItem['id'] ? lastItem['id'] : undefined;
+          newData["id"] = lastId ? lastId + 1: 1;
+        } else {
+          newData["id"] = 1;
+        }
+        newData["timestamp"] = getTimestamp();
+        const jsonNewData = jsonData;
+        jsonNewData.push(newData);
+        await fs.promises.writeFile(this._fileDir,JSON.stringify(jsonNewData));
+        cliSuccess('Object saved!')
+        return await newData['id'];
+      } catch (err: any) {
+        cliError(err['message'])
       }
-      const jsonNewData = jsonData;
-      jsonNewData.push(newData);
-      await fs.promises.writeFile(this._fileDir,JSON.stringify(jsonNewData));
-      cliSuccess('Product saved!')
-      return await newData['id'];
-    } catch (err: any) {
-      cliError(err['message'])
+    } else {
+      return undefined
     }
   }
   /* ---------------- End save ---------------- */
@@ -71,20 +93,24 @@ export default class Container {
     id: number,
     element: object
   ): Promise<any> {
-    try {
-      const data: any = await this.getAll()
-      const elemIndex = data.findIndex((elem: any)=> elem['id'] === id)
-      if (elemIndex){
-        data[elemIndex] = {
-          id: id,
-          ...element
+    if (this._fieldsExist(element)) {
+      try {
+        const data: any = await this.getAll()
+        const elemIndex = data.findIndex((elem: any)=> elem['id'] === id)
+        if (elemIndex){
+          data[elemIndex] = {
+            id: id,
+            ...element
+          }
+          await fs.promises.writeFile(this._fileDir, JSON.stringify(data))
+          cliSuccess(`Object with id ${id} updated!`);
+          return data;
         }
-        await fs.promises.writeFile(this._fileDir, JSON.stringify(data))
-        cliSuccess(`Object with id ${id} updated!`);
-        return data;
+      } catch (err: any) {
+        cliError(err['message']);
       }
-    } catch (err: any) {
-      cliError(err['message']);
+    } else {
+      return undefined
     }
   }
   /* ---------------- End update ---------------- */
